+++ app/Game/Werewolf/Variant/NoRoleKnowledge/Command.hs
@@ -15,10 +15,10 @@ variant.
 
 module Game.Werewolf.Variant.NoRoleKnowledge.Command (
     -- * Ping
-    rolePingedText, werewolvesPingedText,
+    nocturnalRolePingedText, werewolvesPingedText,
 
     -- * Status
-    currentTurnText,
+    currentNocturnalTurnText,
 ) where
 
 import Data.String.Interpolate.Extra
@@ -26,11 +26,11 @@ import Data.Text                     (Text)
 
 import Game.Werewolf
 
-rolePingedText :: Role -> Text
-rolePingedText _ = [iFile|variant/no-role-knowledge/command/ping/role-pinged.txt|]
+nocturnalRolePingedText :: Role -> Text
+nocturnalRolePingedText _ = [iFile|variant/no-role-knowledge/command/ping/nocturnal-role-pinged.txt|]
 
 werewolvesPingedText :: Text
 werewolvesPingedText = [iFile|variant/no-role-knowledge/command/ping/werewolves-pinged.txt|]
 
-currentTurnText :: Game -> Text
-currentTurnText _ = [iFile|variant/no-role-knowledge/command/status/current-turn.txt|]
+currentNocturnalTurnText :: Game -> Text
+currentNocturnalTurnText _ = [iFile|variant/no-role-knowledge/command/status/current-nocturnal-turn.txt|]