+++ app/Game/Werewolf/Variant/Standard/Command.hs
@@ -36,13 +36,15 @@ module Game.Werewolf.Variant.Standard.Command (
     witchCommandsText, witchsTurnText,
 
     -- * Ping
-    playerPingedText, rolePingedText, villagePingedText, werewolvesPingedText,
+    diurnalRolePingedText, nocturnalRolePingedText, playerPingedText, villagePingedText,
+    werewolvesPingedText,
 
     -- * Quit
     callerQuitText,
 
     -- * Status
-    alivePlayersText, currentTurnText, deadPlayersText, gameOverText,
+    alivePlayersText, currentDiurnalTurnText, currentNocturnalTurnText, deadPlayersText,
+    gameOverText,
 
     -- * Unvote
     callerRescindedVoteText,
@@ -166,11 +168,14 @@ witchCommandsText = [iFile|variant/standard/command/help/witch-commands.txt|]
 witchsTurnText :: Text
 witchsTurnText = [iFile|variant/standard/command/help/witchs-turn.txt|]
 
+diurnalRolePingedText :: Role -> Text
+diurnalRolePingedText role = [iFile|variant/standard/command/ping/diurnal-role-pinged.txt|]
+
+nocturnalRolePingedText :: Role -> Text
+nocturnalRolePingedText role = [iFile|variant/standard/command/ping/nocturnal-role-pinged.txt|]
+
 playerPingedText :: Text
 playerPingedText = [iFile|variant/standard/command/ping/player-pinged.txt|]
-
-rolePingedText :: Role -> Text
-rolePingedText role = [iFile|variant/standard/command/ping/role-pinged.txt|]
 
 villagePingedText :: Text
 villagePingedText = [iFile|variant/standard/command/ping/village-pinged.txt|]
@@ -184,8 +189,11 @@ callerQuitText caller = [iFile|variant/standard/command/quit/caller-quit.txt|]
 alivePlayersText :: Game -> Text
 alivePlayersText game = [iFile|variant/standard/command/status/alive-players.txt|]
 
-currentTurnText :: Game -> Text
-currentTurnText game = [iFile|variant/standard/command/status/current-turn.txt|]
+currentDiurnalTurnText :: Game -> Text
+currentDiurnalTurnText game = [iFile|variant/standard/command/status/current-diurnal-turn.txt|]
+
+currentNocturnalTurnText :: Game -> Text
+currentNocturnalTurnText game = [iFile|variant/standard/command/status/current-nocturnal-turn.txt|]
 
 deadPlayersText :: Game -> Text
 deadPlayersText game = [iFile|variant/standard/command/status/dead-players.txt|]